refactor: getTest (admin) controller types refactored
